+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Import `publish` and `next_args` from `p2panda-rs` `api` module [#279](https://github.com/p2panda/aquadoggo/pull/279) `rs`
 - GraphQL replication service gets and verifies new entries and inserts them into the db [#137](https://github.com/p2panda/aquadoggo/pull/137)
 - Dynamically generated GraphQL types and query fields for accessing materialised documents [#141](https://github.com/p2panda/aquadoggo/pull/141) 🥞
 - `validation` and `domain` modules used for publish and next args API [#204](https://github.com/p2panda/aquadoggo/pull/204) 🥞